requirements.txt
# Core LLM / Agent Frameworks
langchain>=1.2.0
langgraph>=1.0.9
langchain-openai>=1.1.0
langchain-huggingface>=0.1.0
langchain-community>=0.4.0

# Model Context Protocol
mcp>=1.26.0

# Vector Store
faiss-cpu==1.7.4

# Database / ORM
sqlalchemy>=2.0.40
psycopg2-binary==2.9.9

# Utilities
numpy>=1.23,<2.0
pydantic>=2.0,<3.0
python-dotenv>=1.0.0
torch==2.2.2
transformers<4.48.0
sentence-transformers<3.4.0
streamlit>=1.54.0